refactor(prompts,tools): dedup Context Awareness; split scaffold + troubleshoot tools
Implements PR-5 of the configuration audit cleanup epic (#143), covering
Wave D (Context Awareness dedup per PR-4 pilot verdict) and Wave F (tool
file splits).

## Wave D — Context Awareness consolidation into AGENTS.md

PR-4's pilot confirmed AGENTS.md IS automatically inherited by every
subagent via opencode's Instruction loader (verified by observing
existing system prompts contain `Instructions from: .../AGENTS.md`
blocks). The original Wave D dichotomy ("CONFIRMED → build.md" vs
"REFUTED → shared file or duplication") missed a third option: promote
shared blocks into AGENTS.md itself. This PR uses that option.

Changes:
  - AGENTS.md gains a new "## Subagent Context Awareness (applies to
    all subagents)" section + a "## Shared Safety Principles (apply to
    all agents)" section. Both are inherited by every agent via the
    Instruction loader.
  - The "## Context Awareness" 7-line block is removed from all 6
    subagents (`agents/{devops,docs,git-ops,ideate,pilot,scribe}/agent.md`)
    and replaced with a single HTML-comment pointer so the structural
    intent remains visible to readers of the agent.md file.
  - For devops, the unique "### Structured Briefs" subsection is
    preserved by being promoted to `## Structured Briefs` (parent
    heading bumped now that the wrapping Context Awareness section is
    gone).

The per-agent "## Safety Rules" / "## Safety Model" sections are
intentionally left in place. They are largely agent-specific (devops
workspace-path forwarding rules, git-ops staging/commit discipline,
pilot's read-access-to-main-project carve-out) and aggressive dedup
risks losing important per-agent nuances. AGENTS.md's new Shared Safety
Principles section captures only the genuinely cross-cutting rules.

## Wave F — Tool file splits

Pilot verification before splitting:

  - `opencode` does NOT load tools from nested directories
    (`tools/<dir>/index.ts` or `tools/<dir>/<file>.ts`). Tools must be
    flat. Confirmed via /tmp/pilot-pr5-tool-nesting/.opencode/tools/foo/
    not appearing in `opencode debug agent build` output.
  - Underscore-prefixed files (e.g., `tools/_helper.ts`) ARE loaded as
    tool files; the underscore convention does not exempt them. So
    "shared helper" files in `tools/` must either (a) export no `tool()`
    calls or (b) be named non-conflictingly. This PR uses approach (a)
    for the scaffold-templates-* files (only export string-template
    generators, no tools) and inline-duplicates the small `run()`
    helper across the troubleshoot-* split files.

### tools/troubleshoot.ts → 3 files

  - troubleshoot-system.ts (316 lines) — check_ports, check_dns,
    check_connectivity, system_info, process_list, disk_usage,
    container_health.
  - troubleshoot-logs.ts (545 lines) — read_logs, extract_errors,
    compare_logs, discover_sources + log-helper functions
    (inferSource, matchesSeverity, extractWithContext, SourceType).
  - troubleshoot-metrics.ts (463 lines) — query_metrics,
    query_metrics_range, discover_prometheus, check_targets +
    Prometheus helpers (findPrometheusEndpoint, parseDuration,
    formatPromResult).
  - The shared `run()` helper (11 lines) is inline-duplicated in all
    three files rather than extracted to avoid creating a fourth
    no-tool helper file.

After split, the original 15 tool functions become 15 tool IDs with
new prefixes: `troubleshoot-system_*` (7), `troubleshoot-logs_*` (4),
`troubleshoot-metrics_*` (4). Verified runtime via
`opencode debug agent build` against a pilot workspace.

### tools/scaffold.ts → 8 files (1 main + 7 template groups)

The original 3249-line file had a `tool()` definition wrapping ~28
template-string-returning generators (Makefile, 4 shell scripts,
Dockerfile, .dockerignore, 3 Cloud Build YAMLs, 7 Terraform files,
config.py, config.toml.example, .env.example, deployer-role YAML, ADR
template, AGENTS.local section, gitignore entries) plus 9 component
scaffolders that orchestrate file writes. Split into:

  - scaffold.ts (364 lines) — imports, ProjectType, detectProject,
    projectLabel, file helpers (ensureDir, safeWrite), component
    scaffolders (scaffoldMakefile, scaffoldScripts, ...,
    scaffoldGitignore), and the `tool()` export. Imports all template
    generators from the 7 split files.
  - scaffold-templates-makefile.ts (119 lines) — generateMakefile.
  - scaffold-templates-scripts.ts (1276 lines) — generateCommonSh,
    generateLocalSh, generateContainerSh, generateCloudSh. (The cloud.sh
    template alone is ~700 lines, so this is unavoidably the biggest
    split file.)
  - scaffold-templates-container.ts (151 lines) — generateDockerfile,
    generateDockerignore.
  - scaffold-templates-cloudbuild.ts (159 lines) — generateCloudbuildYaml,
    generateCloudbuildPlanYaml, generateCloudbuildApplyYaml + 2 helpers
    (_tfEnvBlock, _tfSubstitutionsBlock).
  - scaffold-templates-terraform.ts (567 lines) — generateTfProviders,
    generateTfBackend, generateTfVariables, generateTfMain,
    generateTfOutputs, generateTfLb, generateTfDns.
  - scaffold-templates-iam.ts (77 lines) — generateDeployerRoleYaml.
  - scaffold-templates-docs.ts (581 lines) — generateConfigPy,
    generateConfigTomlExample, generateEnvExample, generateAdrTemplate,
    generateAgentsLocalSection, gitignoreEntries.

`ProjectType` is duplicated as a 1-line union literal in each file that
needs it (makefile, scripts, container, docs) to avoid cross-file type
imports — TypeScript treats identical string-literal unions as
structurally compatible, and opencode plugin loader treats each
tools/*.ts as an independent module.

Net effect: 1 file at 3249 lines → 8 files averaging ~430 lines, with
the dispatch (`scaffold.ts`) now under 400 lines. Diff-review cost on
future scaffold changes drops by an order of magnitude.

### Permission glob updates

After the troubleshoot split, the existing `troubleshoot_*: false`
permission globs in 5 subagents would silently fail to match the new
hyphenated tool IDs (PR-4 pilot confirmed hyphens are preserved).
Updated each to the three new prefixes:

  - agents/{docs,git-ops,ideate,pilot,scribe}/agent.md:
      troubleshoot_*: false      (1 line)
        →
      troubleshoot-logs_*: false
      troubleshoot-metrics_*: false
      troubleshoot-system_*: false  (3 lines)

The devops agent has no `troubleshoot_*: false` glob (it needs the
tools), so no change there.

## Smoke tests (all pass)

  - bash -n install.sh                          → parses cleanly.
  - bun build each new tool file (--external='*') → all 11 build OK.
  - opencode debug agent build (pilot workspace) → all 15 new
    troubleshoot tool IDs exposed with expected prefixes.
  - git grep "agent_workspace_"                 → zero (re-check PR-1).
  - Stale post-#142 cloud verbs                 → zero outside historical
                                                  refs (re-check PR-2).
  - git grep -i antigravity                     → only intentional
                                                  CHANGELOG migration
                                                  notes (re-check PR-3).
  - All 6 subagents have zero "## Context Awareness" sections.
  - AGENTS.md has the new "## Subagent Context Awareness" section.

Closes #148
Refs #143
