+++ agents/docs/agent.md
@@ -12,7 +12,9 @@ tools:
   podman_*: false
   gcloud_*: false
   terraform_*: false
-  troubleshoot_*: false
+  troubleshoot-logs_*: false
+  troubleshoot-metrics_*: false
+  troubleshoot-system_*: false
   devops-preflight_*: false
   branch-cleanup_*: false
   # Disable git-ops tools (handled by delegation)
@@ -60,15 +62,10 @@ You are a documentation assistant that maintains clean, minimalist project
 documentation. You focus exclusively on README.md files. Load the
 `readme-conventions` skill for conventions, templates, and workflow guidance.
 
-## Context Awareness
-
-You are a subagent. You receive ONLY the Task tool prompt -- you have NO
-access to the parent conversation's history. If the prompt contains ambiguous
-references (e.g., "the above docs", "what we discussed"), STOP immediately
-and return a clear message explaining what context is missing. Do NOT guess
--- the parent agent must re-invoke you with a fully self-contained prompt.
-
 ## Core Responsibilities
+
+<!-- Subagent Context Awareness is defined in AGENTS.md and applies here. -->
+
 
 1. **Generate README.md** -- Create a new README.md from scratch based on
    project analysis. Load `readme-conventions` for structure and conventions.
