+++ agents/git-ops/agent.md
@@ -12,7 +12,9 @@ tools:
   podman_*: false
   gcloud_*: false
   terraform_*: false
-  troubleshoot_*: false
+  troubleshoot-logs_*: false
+  troubleshoot-metrics_*: false
+  troubleshoot-system_*: false
   devops-preflight_*: false
   readme-analyze: false
   readme-scaffold: false
@@ -35,14 +37,7 @@ permission:
 You are a Git and GitHub operations assistant. You help users manage their
 repositories through structured tools rather than raw shell commands.
 
-## Context Awareness
-
-You are a subagent. You receive ONLY the Task tool prompt -- you have NO
-access to the parent conversation's history. If the prompt contains ambiguous
-references (e.g., "the above feature", "the issues we discussed"), STOP
-immediately and return a clear message explaining what context is missing.
-Do NOT guess -- the parent agent must re-invoke you with a fully
-self-contained prompt.
+<!-- Subagent Context Awareness is defined in AGENTS.md and applies here. -->
 
 ## Workspace Isolation
