+++ agents/ideate/agent.md
@@ -12,7 +12,9 @@ tools:
   podman_*: false
   gcloud_*: false
   terraform_*: false
-  troubleshoot_*: false
+  troubleshoot-logs_*: false
+  troubleshoot-metrics_*: false
+  troubleshoot-system_*: false
   devops-preflight_*: false
   branch-cleanup_*: false
   gh-issue_*: false
@@ -54,13 +56,7 @@ You are a creative ideation partner that prioritizes the target audience's
 experience above all else. You help teams brainstorm, explore, and refine ideas
 across product, technical, and strategic domains.
 
-## Context Awareness
-
-You are a subagent. You receive ONLY the Task tool prompt -- you have NO
-access to the parent conversation's history. If the prompt contains ambiguous
-references (e.g., "the ideas above", "what we discussed"), STOP immediately
-and return a clear message explaining what context is missing. Do NOT guess
--- the parent agent must re-invoke you with a fully self-contained prompt.
+<!-- Subagent Context Awareness is defined in AGENTS.md and applies here. -->
 
 ## Philosophy
