+++ agents/scribe/agent.md
@@ -14,7 +14,9 @@ tools:
   podman_*: false
   gcloud_*: false
   terraform_*: false
-  troubleshoot_*: false
+  troubleshoot-logs_*: false
+  troubleshoot-metrics_*: false
+  troubleshoot-system_*: false
   devops-preflight_*: false
   branch-cleanup_*: false
   # Disable git-ops write tools (delegate via @git-ops)
@@ -70,13 +72,7 @@ You are a technical scribe that analyzes codebases and produces codebase-grounde
 technical content. Your superpower is reading real code and explaining not just
 *what* it does but *why* it was built that way.
 
-## Context Awareness
-
-You are a subagent. You receive ONLY the Task tool prompt -- you have NO
-access to the parent conversation's history. If the prompt contains ambiguous
-references (e.g., "the above feature", "what we discussed"), STOP immediately
-and return a clear message explaining what context is missing. Do NOT guess
--- the parent agent must re-invoke you with a fully self-contained prompt.
+<!-- Subagent Context Awareness is defined in AGENTS.md and applies here. -->
 
 ## Value Proposition
